improved error handling and code quality

+++ src/controllers/courses.js
@@ -1,5 +1,5 @@
 const {
-  getCourseProvider,
+  getCourseByFieldProvider,
   createCourseProvider,
   updateCourseProvider,
   deleteCourseProvider,
@@ -13,9 +13,9 @@ const {
   createCommaSeperatedKeyValueString
 } = require('../utils');
 
-const {getLessonsByCourseId} = require('./lessons');
-
-const bcrypt = require('bcrypt');
+const {getLessonsByField} = require('./lessons');
+const {BadRequest, Unauthorised} = require('../errors');
+const { Courses, Users, Lessons, ErrorMessages } = require('../constants');
 
 async function insertCourse(course) {
   removeUndefinedProperties(course);
@@ -24,19 +24,19 @@ async function insertCourse(course) {
   keysString = camelToSnakeCase(keysString);
 
   const valuesString = createCommaSeperatedString(Object.keys(course).map(key => course[key]), true);
-  if(!valuesString) return;
+  if(!valuesString) throw new BadRequest(ErrorMessages.MISSING_FIELDS);
   return createCourseProvider(keysString, valuesString);
 }
 
-async function getCoursesByField(fieldValue, getByField='username') {
-  let keysString = createCommaSeperatedString(['course_id, course_name, username, active_lesson_id']);
-  const response = await getCourseProvider(keysString, fieldValue, getByField);
+async function getCoursesByField(getByValue, getByField) {
+  let keysString = createCommaSeperatedString([Courses.COURSE_ID, Courses.COURSE_NAME, Users.USERNAME, Lessons.ACTIVE_LESSON_ID]);
+  const response = await getCourseByFieldProvider(keysString, getByValue, getByField);
   const courses = response.rows;
   return addLessonsToCourses(courses);
 }
 
 async function getAllCourses() {
-  let keysString = createCommaSeperatedString(['course_id, course_name, username, active_lesson_id']);
+  let keysString = createCommaSeperatedString([Courses.COURSE_ID, Courses.COURSE_NAME, Users.USERNAME, Lessons.ACTIVE_LESSON_ID]);
   const response = await getAllCoursesProvider(keysString);
   const courses = response.rows;
   return addLessonsToCourses(courses);
@@ -44,12 +44,17 @@ async function getAllCourses() {
 
 async function addLessonsToCourses(courses) {
   return Promise.all(courses.map(async (course) => {
-    const lessons = await getLessonsByCourseId(course.course_id, true);
+    const lessons = await getLessonsByField(course.course_id, Courses.COURSE_ID, true);
     return {...course, lessons: lessons.rows};
   }));
 }
 
 async function updateCourse(course, courseId) {
+  const response = await getCourseByFieldProvider(Courses.COURSE_ID, courseId, Courses.COURSE_ID);
+  if(!response.rows.length) {
+    throw new BadRequest(`Course doesn't exist`);
+  }
+
   removeUndefinedProperties(course);
 
   Object.keys(course).forEach(key => {
@@ -57,11 +62,20 @@ async function updateCourse(course, courseId) {
   })
 
   let keyValueString = createCommaSeperatedKeyValueString(course, true);
-  if(!keyValueString) return;
+  if(!keyValueString) throw new BadRequest(ErrorMessages.MISSING_FIELDS);;
   return updateCourseProvider(keyValueString, courseId);
 }
 
-function deleteCourse(courseId) {
+async function deleteCourse(courseId, loggedInUser) {
+  const response = await getCourseByFieldProvider(Courses.COURSE_ID, courseId, Courses.COURSE_ID);
+  if(!response.rows.length) {
+    throw new BadRequest(`Course doesn't exist`);
+  }
+
+  const courses = await getCoursesByField(courseId, Courses.COURSE_ID);
+  if(courses[0]?.username != loggedInUser) {
+    throw new Unauthorised('Unauthorised user');
+  }
   return deleteCourseProvider(courseId);
 }
